+++ itsystems/admin.py
@@ -1,30 +1,30 @@
 from django.contrib import admin
 
-from .models import SystemInstance, HostInstance, AgentService, Agent, Vendor, Component
+from .models import System, Host, AgentService, Agent, Vendor, Component
 
 
-class HostInstancesInLine(admin.TabularInline):
-    model = HostInstance
+class HostsInLine(admin.TabularInline):
+    model = Host
 
 class AgentInLine(admin.TabularInline):
     model = Agent
 
-class SystemInstanceAdmin(admin.ModelAdmin):
+class SystemAdmin(admin.ModelAdmin):
     ordering = ('name', 'sdlc_stage')
     list_display = ('name', 'sdlc_stage', 'id')
     inlines = [
-        HostInstancesInLine,
+        HostsInLine,
     ]
 
-class HostInstanceAdmin(admin.ModelAdmin):
+class HostAdmin(admin.ModelAdmin):
     ordering = ('name',)
-    list_display = ('name', 'system_instance', 'host_type', 'os', 'id')
+    list_display = ('name', 'system', 'host_type', 'os', 'id')
     inlines = [
         AgentInLine,
     ]
 
 class AgentAdmin(admin.ModelAdmin):
-    list_display = ('agent_id', 'host_instance', 'agent_service')
+    list_display = ('agent_id', 'host', 'agent_service')
 
 class VendorAdmin(admin.ModelAdmin):
     list_display = ('name',)
@@ -32,8 +32,8 @@ class VendorAdmin(admin.ModelAdmin):
 class ComponentAdmin(admin.ModelAdmin):
     list_display = ('name', 'version')
 
-admin.site.register(SystemInstance, SystemInstanceAdmin)
-admin.site.register(HostInstance, HostInstanceAdmin)
+admin.site.register(System, SystemAdmin)
+admin.site.register(Host, HostAdmin)
 admin.site.register(Vendor, VendorAdmin)
 admin.site.register(Component, ComponentAdmin)
 admin.site.register(AgentService)